+++ src/rotp/ui/races/RacesStatusUI.java
@@ -690,7 +690,7 @@ public final class RacesStatusUI extends BasePanel implements MouseListener, Mou
                 int ptX = startX+(displayW*i/totalTurns);
                 if ((ptX - prevX) > maxPtSpacing)
                     ptX = prevX + maxPtSpacing;
-                int ptY = maxYValue == 0 ? 0 : startY-(int)((float)displayH*playerHigh[i]/maxYValue);
+                int ptY = maxYValue == 0 ? 0 : startY-(int)(displayH*playerHigh[i]/maxYValue);
                 if (prevY >= 0)
                     g.drawLine(prevX, prevY, ptX, ptY);
                 prevX = ptX;
@@ -726,7 +726,7 @@ public final class RacesStatusUI extends BasePanel implements MouseListener, Mou
             int ptX = startX+(displayW*i/totalTurns);
             if ((ptX - prevX) > maxPtSpacing)
                 ptX = prevX + maxPtSpacing;
-            int ptY = maxYValue == 0 ? 0 : startY-(int)((float)displayH*playerVals[i]/maxYValue);
+            int ptY = maxYValue == 0 ? 0 : startY-(int)(displayH*playerVals[i]/maxYValue);
             if (prevY >= 0)
                 g.drawLine(prevX, prevY, ptX, ptY);
             prevX = ptX;
@@ -749,7 +749,7 @@ public final class RacesStatusUI extends BasePanel implements MouseListener, Mou
             int ptX = startX+(displayW*i/totalTurns);
             if ((ptX - prevX) > maxPtSpacing)
                 ptX = prevX + maxPtSpacing;
-            int ptY = maxYValue == 0 ? 0 : startY-(int)((float)displayH*empireVals[i]/maxYValue);
+            int ptY = maxYValue == 0 ? 0 : startY-(int)(displayH*empireVals[i]/maxYValue);
             if (prevY >= 0)
                 g.drawLine(prevX, prevY, ptX, ptY);
             prevX = ptX;
